Fixed The Project for Assignment6 and Assignment6.Tests
