+++ Assignment6.Tests/ArrayTests.cs
@@ -1,0 +1,117 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using System.Collections.Generic;
+
+namespace Assignment6.Tests
+{
+    [TestClass]
+    public class ArrayTests
+    {
+        [DataTestMethod]
+        [DataRow(1)]
+        [DataRow(2)]
+        [DataRow(3)]
+        public void Array_Add_Success(int item)
+        {
+            Array<int> array = new Array<int>(3)
+            {
+                item
+            };
+
+            Assert.IsTrue(array.Contains(item));
+        }
+
+        [TestMethod]
+        public void Array_Clear_Success()
+        {
+            Array<int> array = new Array<int>(2)
+            {
+                1,
+                2
+            };
+            array.Clear();
+
+            Assert.IsTrue(array.Count == 0);
+        }
+
+        [DataTestMethod]
+        [DataRow(1)]
+        [DataRow(2)]
+        [DataRow(3)]
+        [DataRow(4)]
+        [DataRow(5)]
+        public void Array_Contains_ReturnsItem(int item)
+        {
+            Array<int> array = new Array<int>(5)
+            {
+                1,
+                2,
+                3,
+                4,
+                5
+            };
+
+            Assert.IsTrue(array.Contains(item));
+        }
+
+        [DataTestMethod]
+        [DataRow(6)]
+        [DataRow(7)]
+        [DataRow(8)]
+        [DataRow(9)]
+        [DataRow(10)]
+        public void Array_Contains_DoesNotContain(int item)
+        {
+            Array<int> array = new Array<int>(5)
+            {
+                1,
+                2,
+                3,
+                4,
+                5
+            };
+
+            Assert.IsFalse(array.Contains(item));
+        }
+
+        [DataTestMethod]
+        [DataRow(1)]
+        [DataRow(2)]
+        [DataRow(3)]
+        [DataRow(4)]
+        [DataRow(5)]
+        public void Array_Remove_Success(int item)
+        {
+            Array<int> array = new Array<int>(5)
+            {
+                1,
+                2,
+                3,
+                4,
+                5
+            };
+
+            Assert.IsTrue(array.Remove(item));
+        }
+
+        [DataTestMethod]
+        [DataRow(6)]
+        [DataRow(7)]
+        [DataRow(8)]
+        [DataRow(9)]
+        [DataRow(10)]
+        public void Array_Removes_DoesNotContain(int item)
+        {
+            Array<int> array = new Array<int>(5)
+            {
+                1,
+                2,
+                3,
+                4,
+                5
+            };
+
+            Assert.IsFalse(array.Remove(item));
+        }
+    }
+}
