+++ Assignment6.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.29418.71
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment6", "Assignment6\Assignment6.csproj", "{4C7EC39E-FDB6-4D15-89DF-15AB0BADF554}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assignment6.Tests", "Assignment6.Tests\Assignment6.Tests.csproj", "{7D152D79-5589-414B-A38E-84CAA41EE6FC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -39,6 +41,26 @@ Global
 		{4C7EC39E-FDB6-4D15-89DF-15AB0BADF554}.Release|x64.Build.0 = Release|Any CPU
 		{4C7EC39E-FDB6-4D15-89DF-15AB0BADF554}.Release|x86.ActiveCfg = Release|Any CPU
 		{4C7EC39E-FDB6-4D15-89DF-15AB0BADF554}.Release|x86.Build.0 = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|ARM.Build.0 = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|x64.Build.0 = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Debug|x86.Build.0 = Debug|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|ARM.ActiveCfg = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|ARM.Build.0 = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|ARM64.Build.0 = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|x64.ActiveCfg = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|x64.Build.0 = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|x86.ActiveCfg = Release|Any CPU
+		{7D152D79-5589-414B-A38E-84CAA41EE6FC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
