+++ Assignment6/Assignment6.csproj
@@ -7,8 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MSTest.TestFramework" Version="2.0.0" />
-    <PackageReference Include="NUnit" Version="3.12.0" />
   </ItemGroup>
 
 </Project>
